requirements.txt
nose>=1.1.2
nose_parameterized
scikit-learn>=0.17
Theano>=0.8
Lasagne>=0.1
sklearn-contrib-lightning
